+++ philo/src/print_test.c
@@ -6,7 +6,7 @@
 /*   By: gaducurt <gaducurt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/16 11:38:59 by gaducurt          #+#    #+#             */
-/*   Updated: 2025/10/01 10:50:42 by gaducurt         ###   ########.fr       */
+/*   Updated: 2025/10/02 17:07:45 by gaducurt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,27 +15,25 @@
 
 void	print_tab_fork(t_fork *tab_fork, t_common common)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i <= common.nb_philo - 1)
 	{
-		printf("tab_fork[%d].id_fork = %d\ntab_fork[%d].avalable = %d\n", i, tab_fork[i].id_fork, i, tab_fork[i].avalable.data);
+		printf("tab_fork[%zu].id_fork = %d\ntab_fork[%zu].avalable = %zu\n", i, tab_fork[i].id_fork, i, tab_fork[i].avalable.data);
 		i++;
 	}
 }
 
 void	print_tab_philo(t_philo *tab_philo, t_common common)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i <= common.nb_philo - 1)
 	{
-		printf("tab_philo[%d].philo_id = %d\n", i, tab_philo[i].philo_id);
-		printf("common.nb_philo = %d\n", common.nb_philo);
-		// printf("tab_philo[%d].left_fork = %d\n", i, tab_philo[i].left_fork->id_fork);
-		// printf("tab_philo[%d].right_fork = %d\n", i, tab_philo[i].right_fork->id_fork);
+		printf("tab_philo[%zu].philo_id = %d\n", i, tab_philo[i].philo_id);
+		printf("common.nb_philo = %zu\n", common.nb_philo);
 		printf("\n");
 		i++;
 	}
